refactor: rename react directory for users module
